Multiplicative_Functions class

+++ formatting.cpp
@@ -2,7 +2,7 @@
 #include <stdexcept>
 #include <iomanip>
 
-void check_length (const size_t actual_size, const int EXPECTED_SIZE) {
+void check_length (const size_t actual_size, const size_t EXPECTED_SIZE) {
     if (actual_size != EXPECTED_SIZE) {
         throw std::invalid_argument("Error: Vector length does not match expected input.");
     }
@@ -15,7 +15,7 @@ void generate_header(const std::vector<std::string>& HEADER_LIST) {
     }
     std::cout << "\n";}
 
-void generate_row(const std::vector<int>& row_data, const int VECTOR_SIZE) {
+void generate_row(const std::vector<int>& row_data, const size_t VECTOR_SIZE) {
     check_length(row_data.size(), VECTOR_SIZE);
     std::cout << "|";
     for (int number : row_data) {
